move code outside of datagrid
