+++ client/src/components/forms/DataGrid.tsx
@@ -1,13 +1,9 @@
-import client from '@/client'
-import { faFileExport } from '@fortawesome/free-solid-svg-icons'
 import {
   FontAwesomeIcon,
   FontAwesomeIconProps,
 } from '@fortawesome/react-fontawesome'
 import { Button, ButtonProps } from '@heroui/button'
-import { addToast } from '@heroui/react'
-import React, { useCallback } from 'react'
-import { useTranslation } from 'react-i18next'
+import React from 'react'
 import { EmptyState } from '../table/EmptyState'
 
 export function Titlebar({ title }: { title: string }) {
@@ -35,77 +31,17 @@ export function FilterButton({ icon, title, ...props }: FilterButtonProps) {
   )
 }
 
-export const SelectableContext = React.createContext<{
-  selectable: boolean
-  toggleSelectable: () => void
-  selected: string[]
-  setSelected: React.Dispatch<React.SetStateAction<string[]>>
-}>({
-  selectable: false,
-  toggleSelectable: () => {},
-  selected: [],
-  setSelected: () => {},
-})
-
-function useExportProductTree() {
-  const { t } = useTranslation()
-
-  return client.useMutation('get', '/api/v1/products/export', {
-    onSuccess: (response) => {
-      const blob = new Blob([JSON.stringify(response, null, 2)], {
-        type: 'application/json',
-      })
-
-      const url = URL.createObjectURL(blob)
-      const a = document.createElement('a')
-      a.href = url
-      a.download = `product_tree_export_${Date.now()}.json`
-      document.body.appendChild(a)
-      a.click()
-      a.remove()
-      URL.revokeObjectURL(url)
-    },
-    onError: (error) => {
-      addToast({
-        title: t('export.error.title'),
-        description: error?.title || t('export.error.text'),
-        color: 'danger',
-      })
-    },
-  })
-}
-
 export default function DataGrid({
   title,
   actions,
-  exportable = false,
   addButton,
   children,
 }: {
   title?: string
   actions?: React.ReactNode
-  exportable?: boolean
   addButton?: React.ReactNode
   children: React.ReactNode[] | React.ReactNode | undefined
 }) {
-  const { t } = useTranslation()
-  const [selected, setSelected] = React.useState<string[]>([])
-  const [selectable, setSelectable] = React.useState<boolean>(false)
-  const toggleSelectable = () => {
-    setSelectable(!selectable)
-    setSelected([])
-  }
-
-  const exportMutation = useExportProductTree()
-
-  const onExportClick = useCallback(() => {
-    exportMutation.mutate({
-      params: {
-        query: { ids: selected.join(',') },
-      },
-    })
-  }, [exportMutation, selected])
-
   return (
     <div className="flex w-full flex-col items-center gap-4">
       {title && (
@@ -116,48 +52,11 @@ export default function DataGrid({
         </div>
       )}
 
-      {exportable && (
-        <div className="flex w-full items-center justify-end gap-2">
-          {selectable ? (
-            <>
-              <Button variant="light" color="danger" onPress={toggleSelectable}>
-                {t('export.stopSelection')}
-              </Button>
-
-              <Button
-                color="primary"
-                onPress={onExportClick}
-                isDisabled={selected.length === 0}
-              >
-                <FontAwesomeIcon icon={faFileExport} />
-                {t('export.exportSelected', { count: selected.length })}
-              </Button>
-            </>
-          ) : (
-            <Button variant="light" color="primary" onPress={toggleSelectable}>
-              {t('export.label')}
-            </Button>
-          )}
-        </div>
-      )}
-
       {React.Children.count(children) === 0 ? (
         <EmptyState add={addButton} />
       ) : null}
 
-      <SelectableContext.Provider
-        value={{
-          selectable,
-          toggleSelectable: () => {
-            setSelectable(!selectable)
-            setSelected([])
-          },
-          selected,
-          setSelected,
-        }}
-      >
-        <div className="flex w-full flex-col gap-2">{children}</div>
-      </SelectableContext.Provider>
+      <div className="flex w-full flex-col gap-2">{children}</div>
 
       {/* <Pagination /> */}
     </div>
